Add barebones inference
Motivation:
Implement basic sentiment analysis to the main view.

Modifications:
Added a viewmodel to the main view with a function to perform a
sentiment analysis on a paragraph. This functions returns a string
with the results of the analysis using a enum called `Sentiment`.

Result:
Basis sentiment analysis can now be performed on the main view.