Add NavBar and TabBar to view hierarchy
Motivation:

Our prototype determines that there will be at least 2 more screens to
the project. With that in mind, we need to place all the views inside
both a Navigation Controller and Tab Bar Controller.
We decided to do already in the begining of the app to prevent further
problems.

Modifications:

- Placed `MainViewController` inside a `UINavigationViewController` in
  `SceneDelegate.swift`.
- Placed `UINavigationViewController` inside a `UITabBarController` in
  `SceneDelegate.swift`.
- Changed `MainViewController` title to "Coremelysis".

Result:

Main screen now displays both a navbar with the Coremelysis and a
tabBar.