SwiftLint and ML Manager added to project
Motivation:
SwiftLint yml file was already present in the project but we forgot to
add the script to the project build phase.
Machine Learning needed to be separeted into a different class for
better encapsulation and to facilitate the exchange of models for
inference.

Modifications:
- Machine Learning was separated into a manager class.
- Script was added to the build phase to run SwiftLint.
- Several other changes in multiple files to conform to SwiftLint, all
  of them "cosmetic" like whitespace removals and line length
shortenings.

Result:
SwiftLint now runs every time project is build ensuring some level of
code quality.
Machine learning can now be invoked from an object and has the initial
preparations to run different types of model.