Add Alternative Sentiment Analysis Model
Motivation:

To prevent bias and offer counterproof to the inferences, we sought to
add a new model for Sentiment Analysis.
This model was found on github and can only perform sentiment analysis
in english bodies of text but was provided under MIT license and was
used as a proof of concept for other models implementations.

Modifications:

- Added SentimentPolarity.mlmodel to Resources
- Modified MLManager to operate with enum instead of MLModels

Result:

MLManager now can perform a new type of inference named Sentiment
Polarity, given the correct option in the init.