Add Error Handling and Refactoring
Motivation:

Machine Learning is one of the core concepts of the app and as such
should be well documented and handled, as the entire apps experience
depends on it.
MLManager didn't had any kind of state and it made no sense to see it as
an object.

Modifications:

- MLManager class restructured to use static methods.
- Added Machine Learning Error enum file with descriptions for each of
  the cases.
- Added Sentiment Analysis Model enum file for each of the predicted models
  for the app, including custom models, and moved the sentiment ranges
  to this file.
- Made slight adjustments to MainView ViewModel and SceneDelegate to
  conform to the refactoring.

Result:

Errors in the machine learning section of the app now get propagated to
the class consuming it, with more descriptions detailing the causes of
the error. It's now possible to call upon MLManager analyse method
without the need to instantiate an object.