Add UIColor extension
Motivation:
The app should display the correct colors based on the style guide.
In order to avoid having to repeatedly input the same RGB values, a
more convinient way should be implemented.

Modification:
An UIColor extension has been added in which we can define a color
to be used in the codebase using the dot notation.
I've also added a group called Extensions to better organize any and all
extensions we may need.

Result:
Developers can now use the colors determined in the style guide using
dot notation.