Add a custom UITabBarController
Motivation:
The process of initialization in the SceneDelegate was too loaded and
prone to mistakes.

Modification:
A new type has been added, CoremelysisTabBarController. This type is
responsible for the initialization of the UITabBarController used in
the app's navigation.

I've also fixed a minor bug in which the title of each screen's
UITabBarItem would only be set once the screen was loaded.

Result:
A new customizad UITabBarController has been added improving readablity
and maintainability of the SceneDelegate.