Fix HistorySummaryViewController
Motivation:
Using constraints to represent the amount of each type of entries was
resulting in complex and bug prone code, without being good
representations of the data being displayed.

Modification:
UILabels are now being used to display each type (Positive, Negative and
Neutral) of entry.

Result:
The HistorySummaryViewController now has labels showing the total number
of entries and the number of positive, negative and neutral entries.
Also, the implementation is clearer and simpler.